Transition implementation to hidden type. Public API now abstract class.
This facilitates writing test, mock, or delegating implementations of MixpanelAPI. When new APIs are added, we can be assured (via javac and compilation failures) that its impossible miss overriding a method and accidentally call through to a real impl.
